Update README.md
The focus of this research proposal is to address the research brief given by Nespresso. Their main 
business objectives are: 
• Maintain its position as a leader in the premium coffee market. 
• Prioritize product innovation, developing new coffee blends and brewing systems. 
• Increase customer satisfaction. 
• Customers should have positive perceptions of the sustainability of its pods. 
This proposal aims to give an understanding of the business context by examining the coffee market 
in general and Nespresso’s contribution to the coffee market. This proposal is to help Nespresso gain 
insights from its customers and non-customers regarding their maximum willingness to pay, their 
future preferences to identify evolving trends in the consumer appetite, and perception they have 
towards Nespresso and other leading coffee firms. 
An analytical model has been developed with research questions and hypotheses to get a clear 
understanding of the aspects we need to focus on to meet the research aims. The primary data for 
the research will be collected via a questionnaire, which will be generated through online surveys 
and computer assisted personal interviews (CAPI). 
Moreover, the research design comprises of both exploratory and conclusive research. Two different 
sampling frames will be used for customers and non-customers. Sampling techniques such as 
stratified, and convenience sampling will be utilized for customers and non-customers respectively.  
A sample size of 9,260 is initially needed for customers and 7,452 is initially needed for non
customers. Moreover, the data is analyzed using cross-tabulation and one-way ANOVA.